feat: functions for autonomi release information
The `AntReleaseRepoActions` trait is extended with two new public functions for obtaining the
information from releases in the `maidsafe/autonomi` repository.

They return the name of the release, the commit hash, and all the binary information, including the
sha256 hashes and Semantic Version numbers.

This information is going to be used in the node's automatic upgrading process.